clean up; fix clippy
