+++ bi-kzg/src/lib.rs
@@ -1,16 +1,18 @@
 mod bi_fft;
 mod coeff_form_bi_kzg;
-// mod lagrange_form_bi_kzg;
 mod pcs;
 mod poly;
 mod structs;
 mod util;
 
+// mod lagrange_form_bi_kzg;
+
 #[cfg(test)]
 mod tests;
 
 pub use coeff_form_bi_kzg::CoeffFormBiKZG;
-// pub use lagrange_form_bi_kzg::LagrangeFormBiKZG;
 pub use pcs::PolynomialCommitmentScheme;
 pub use structs::BivariatePolynomial;
 pub use structs::{BiKZGCommitment, BiKZGProof, BiKZGSRS, BiKZGVerifierParam};
+
+// pub use lagrange_form_bi_kzg::LagrangeFormBiKZG;
