+++ bi-kzg/src/structs.rs
@@ -1,6 +1,4 @@
-use std::io::{self, Read, Write};
-
-use halo2curves::{pairing::Engine, serde::SerdeObject};
+use halo2curves::pairing::Engine;
 
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct BivariatePolynomial<F> {
@@ -35,13 +33,6 @@ pub struct BiKZGSRS<E: Engine> {
     /// tau_1 times the above generator of G2.
     pub tau_1_h: E::G2Affine,
 }
-
-// /// `BiKZGProverParam` is used to generate a proof
-// #[derive(Clone, Debug, Eq, PartialEq, Default)]
-// pub struct BiKZGProverParam<E: Engine> {
-//     /// Parameters
-//     pub powers_of_g: Vec<E::G1Affine>,
-// }
 
 /// `UnivariateVerifierParam` is used to check evaluation proofs for a given
 /// commitment.
@@ -80,73 +71,5 @@ impl<E: Engine> From<&BiKZGSRS<E>> for BiKZGVerifierParam<E> {
             tau_0_h: srs.tau_0_h,
             tau_1_h: srs.tau_1_h,
         }
-    }
-}
-
-impl<E: Engine> SerdeObject for BiKZGCommitment<E> {
-    /// The purpose of unchecked functions is to read the internal memory representation
-    /// of a type from bytes as quickly as possible. No sanitization checks are performed
-    /// to ensure the bytes represent a valid object. As such this function should only be
-    /// used internally as an extension of machine memory. It should not be used to deserialize
-    /// externally provided data.
-    fn from_raw_bytes_unchecked(_bytes: &[u8]) -> Self {
-        todo!("Implement this function")
-    }
-    fn from_raw_bytes(_bytes: &[u8]) -> Option<Self> {
-        todo!("Implement this function")
-    }
-
-    fn to_raw_bytes(&self) -> Vec<u8> {
-        todo!("Implement this function")
-    }
-
-    /// The purpose of unchecked functions is to read the internal memory representation
-    /// of a type from disk as quickly as possible. No sanitization checks are performed
-    /// to ensure the bytes represent a valid object. This function should only be used
-    /// internally when some machine state cannot be kept in memory (e.g., between runs)
-    /// and needs to be reloaded as quickly as possible.
-    fn read_raw_unchecked<R: Read>(_reader: &mut R) -> Self {
-        todo!("Implement this function")
-    }
-    fn read_raw<R: Read>(_reader: &mut R) -> io::Result<Self> {
-        todo!("Implement this function")
-    }
-
-    fn write_raw<W: Write>(&self, _writer: &mut W) -> io::Result<()> {
-        todo!("Implement this function")
-    }
-}
-
-impl<E: Engine> SerdeObject for BiKZGProof<E> {
-    /// The purpose of unchecked functions is to read the internal memory representation
-    /// of a type from bytes as quickly as possible. No sanitization checks are performed
-    /// to ensure the bytes represent a valid object. As such this function should only be
-    /// used internally as an extension of machine memory. It should not be used to deserialize
-    /// externally provided data.
-    fn from_raw_bytes_unchecked(_bytes: &[u8]) -> Self {
-        todo!("Implement this function")
-    }
-    fn from_raw_bytes(_bytes: &[u8]) -> Option<Self> {
-        todo!("Implement this function")
-    }
-
-    fn to_raw_bytes(&self) -> Vec<u8> {
-        todo!("Implement this function")
-    }
-
-    /// The purpose of unchecked functions is to read the internal memory representation
-    /// of a type from disk as quickly as possible. No sanitization checks are performed
-    /// to ensure the bytes represent a valid object. This function should only be used
-    /// internally when some machine state cannot be kept in memory (e.g., between runs)
-    /// and needs to be reloaded as quickly as possible.
-    fn read_raw_unchecked<R: Read>(_reader: &mut R) -> Self {
-        todo!("Implement this function")
-    }
-    fn read_raw<R: Read>(_reader: &mut R) -> io::Result<Self> {
-        todo!("Implement this function")
-    }
-
-    fn write_raw<W: Write>(&self, _writer: &mut W) -> io::Result<()> {
-        todo!("Implement this function")
     }
 }
